wire_e2e_identity::acme

Struct RustyAcme

pub struct RustyAcme;

Implementations

impl RustyAcme

pub fn new_account_request( directory: &AcmeDirectory, alg: JwsAlgorithm, kp: &Pem, previous_nonce: String, ) -> RustyAcmeResult<AcmeJws>

Create a new acme account see RFC 8555 Section 7.3

pub fn new_account_response(response: Value) -> RustyAcmeResult<AcmeAccount>

parse the response from POST /acme/new-account see RFC 8555 Section 7.3

impl RustyAcme

pub fn new_authz_request( url: &Url, account: &AcmeAccount, alg: JwsAlgorithm, kp: &Pem, previous_nonce: String, ) -> RustyAcmeResult<AcmeJws>

create authorizations see RFC 8555 Section 7.5

pub fn new_authz_response(response: Value) -> RustyAcmeResult<AcmeAuthz>

parse the response from POST /acme/authz/{authz_id} RFC 8555 Section 7.5

impl RustyAcme

pub fn certificate_req( finalize: AcmeFinalize, account: AcmeAccount, alg: JwsAlgorithm, kp: &Pem, previous_nonce: String, ) -> RustyAcmeResult<AcmeJws>

For fetching the generated certificate see RFC 8555 Section 7.4.2

pub fn certificate_response( response: String, order: AcmeOrder, hash_alg: HashAlgorithm, env: Option<&PkiEnvironment>, ) -> RustyAcmeResult<Vec<Vec<u8>>>

see RFC 8555 Section 7.4.2

impl RustyAcme

pub fn dpop_chall_request( access_token: String, dpop_chall: AcmeChallenge, account: &AcmeAccount, alg: JwsAlgorithm, kp: &Pem, previous_nonce: String, ) -> RustyAcmeResult<AcmeJws>

client id challenge request to POST /acme/challenge/{token} see RFC 8555 Section 7.5.1

pub fn oidc_chall_request( id_token: String, oidc_chall: AcmeChallenge, account: &AcmeAccount, alg: JwsAlgorithm, kp: &Pem, previous_nonce: String, ) -> RustyAcmeResult<AcmeJws>

oidc challenge request to POST /acme/challenge/{token} see RFC 8555 Section 7.5.1

pub fn new_chall_response(response: Value) -> RustyAcmeResult<AcmeChallenge>

parse the response from POST /acme/challenge/{token} RFC 8555 Section 7.5.1

impl RustyAcme

pub fn acme_directory_response( response: Value, ) -> RustyAcmeResult<AcmeDirectory>

First, call the directory endpoint GET /acme/{provisioner_name}/directory. Then pass the response to this method to deserialize it see RFC 8555 Section 7.1.1

impl RustyAcme

pub fn finalize_req( order: &AcmeOrder, account: &AcmeAccount, alg: JwsAlgorithm, acme_kp: &Pem, signing_kp: &Pem, previous_nonce: String, ) -> RustyAcmeResult<AcmeJws>

see RFC 8555 Section 7.4

pub fn finalize_response(response: Value) -> RustyAcmeResult<AcmeFinalize>

see RFC 8555 Section 7.4

impl RustyAcme

pub fn new_order_request( display_name: &str, client_id: ClientId, handle: &Handle, expiry: Duration, directory: &AcmeDirectory, account: &AcmeAccount, alg: JwsAlgorithm, kp: &Pem, previous_nonce: String, ) -> RustyAcmeResult<AcmeJws>

create a new order see RFC 8555 Section 7.4.

pub fn new_order_response(response: Value) -> RustyAcmeResult<AcmeOrder>

parse response from order creation RFC 8555 Section 7.4

impl RustyAcme

pub fn check_order_request( order_url: Url, account: &AcmeAccount, alg: JwsAlgorithm, kp: &Pem, previous_nonce: String, ) -> RustyAcmeResult<AcmeJws>

check an order status until it becomes ready see RFC 8555 Section 7.4

pub fn check_order_response(response: Value) -> RustyAcmeResult<AcmeOrder>

parse response from order check see RFC 8555 Section 7.4