wire_e2e_identity

Module prelude

Re-exports

pub use super::RustyE2eIdentity;

pub use crate::acme::prelude::AcmeDirectory;

pub use crate::acme::prelude::RustyAcme;

pub use crate::acme::prelude::RustyAcmeError;

pub use crate::acme::prelude::WireIdentity;

pub use crate::acme::prelude::WireIdentityReader;

pub use crate::acme::prelude::compute_raw_key_thumbprint;

pub use crate::acme::prelude::x509;

pub use crate::acme::prelude::x509::IdentityStatus;

Structs

E2eiAcmeAccount

E2eiAcmeChallenge

E2eiAcmeFinalize

E2eiAcmeOrder

E2eiClientId

Unique user handle

E2eiNewAcmeOrder

Handle

A unique human-friendly identifier for a user e.g. beltram_wire

Enums

E2eIdentityError

All e2e identity related errors

E2eiAcmeAuthorization

HashAlgorithm

Narrows the supported hashing algorithms to the ones we define

JwsAlgorithm

Narrows the supported signature algorithms to the ones we define

RustyJwtError

All errors which [crate::RustyJwtTools] might throw

Functions

generate_jwk

Generates a json serialized JWK for testing purposes

parse_json_jwk

Parses a raw JWK Json serialized

Type Aliases

E2eIdentityResult

Wrapper over a Result with a RustyAcmeError error