wire_e2e_identity::x509_check::revocation

Struct PkiEnvironment

pub struct PkiEnvironment { /* private fields */ }

Implementations

impl PkiEnvironment

pub fn decode_pem_cert(pem: String) -> RustyX509CheckResult<Certificate>

pub fn decode_der_crl(crl_der: Vec<u8>) -> RustyX509CheckResult<CertificateList>

pub fn extract_ski_aki_from_cert( cert: &Certificate, ) -> RustyX509CheckResult<(String, Option<String>)>

pub fn encode_cert_to_der(cert: &Certificate) -> RustyX509CheckResult<Vec<u8>>

pub fn encode_crl_to_der(crl: &CertificateList) -> RustyX509CheckResult<Vec<u8>>

pub fn init( params: PkiEnvironmentParams<'_>, ) -> RustyX509CheckResult<PkiEnvironment>

Initializes a certval PkiEnvironment using the provided params

pub fn set_time_of_interest(&mut self, toi: u64)

Overrides TIME_OF_INTEREST for certificate verifications based on a moment in the past or future

pub fn refresh_time_of_interest(&mut self) -> RustyX509CheckResult<()>

Updates the TIME_OF_INTEREST for certificate checks to be now

pub fn validate_trust_anchor_cert( &self, cert: &Certificate, ) -> RustyX509CheckResult<()>

pub fn validate_crl(&self, crl: &CertificateList) -> RustyX509CheckResult<()>

👎Deprecated: This method is not to be used as it causes spurious verification failures because of re-encoding the DER repr of the CRL. Use validate_crl_with_raw

pub fn validate_crl_with_raw( &self, crl_raw: &[u8], ) -> RustyX509CheckResult<CertificateList>

pub fn validate_cert( &self, end_identity_cert: &Certificate, ) -> RustyX509CheckResult<()>

pub fn validate_cert_and_revocation( &self, end_identity_cert: &Certificate, ) -> RustyX509CheckResult<()>